add createUser

+++ main-service/src/main/java/ru/practicum/model/Users.java
@@ -14,13 +14,10 @@ import java.util.Objects;
 @Entity
 @Table(name = "users")
 public class Users {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private Integer id;
     private String name;
-
     private String email;
 
     @Override
